requirements.txt
fastapi>=0.105.0
uvicorn>=0.24.0
langchain-community>=0.2.0
langchain-core>=0.2.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pydantic>=1.10.13
requests>=2.31.0
